added rain color

+++ src/main/java/com/app/ping/weather/WeatherReport.java
@@ -1,6 +1,6 @@
 package com.app.ping.weather;
 
-public record WeatherReport(int clouds, double rain)
+public record WeatherReport(int clouds, int rain)
 {
 
 }
